+++ src/api/mod.rs
@@ -1,0 +1,3 @@
+pub mod schema;
+pub mod core_instruction_handler;
+pub mod plugin_instruction_handler;
