+++ src/lib.rs
@@ -1,3 +1,1 @@
 pub mod api;
-pub mod core_instruction_handler;
-pub mod plugin_instruction_handler;
